[LayoutNG] Do not examine sibling nodes when walking break tokens.
When processing incoming break tokens as part of block child layout, do
not call NGLayoutInputNode::NextSibling(), because:

1: We're not going to use the result
2: It would trigger a DCHECK failure in some cases

If we're resuming at a float in an inline formatting context, we'll do
this directly from the block layout algorithm. However, the float may
have inline-level siblings, and those should be handled by the inline
layout algorithm, not the block layout algorithm.
NGBlockNode::NextSibling() (rightly) chokes on inline-level boxes, so
just avoid it. There will always be an inline break token which will
take care of this correctly for us.

The attached test used to DCHECK-fail (but otherwise pass) without this
fix.

Change-Id: I826c30b362f2ae40af12f1b7cf1b0a9e8cd9ccfb