[LayoutNG] Carry over incoming inline break token when no new lines.
If we want to break before the first line in a fragment, we need to
check if there's an incoming inline break token, and add that one to the
builder, rather than assuming that we should always resume at the
beginning of the inline formatting context.

The new test would either freeze or trigger a DCHECK failure without
this fix.

Bug: 1122901
Change-Id: Ib698d477aef18d11dbe1436eaf12d5d52bdc023d